02-MobileNet/ifo-app/requirements.txt
# -f https://download.pytorch.org/whl/torch_stable.html
https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.7.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
streamlit>=0.64.0
# torch>=1.6.0+cpu
# torchvision>=0.7.0+cpu
pandas>=1.1.0
pillow>=7.2.0
googledrivedownloader==0.4
